8pp

+++ Medicine.py
@@ -1,9 +1,9 @@
 from flask import Blueprint, Response, request, jsonify
 from marshmallow import ValidationError
 from flask_bcrypt import Bcrypt
-from shm_metods.model import Medicine, Session
-from shm_metods.shm import MedicineSchema
-
+from model import Medicine, Session, Order, User, Medicine, Session
+from shm import MedicineSchema
+from User import  auth
 medicine = Blueprint('medicine', __name__)
 bcrypt = Bcrypt()
 
@@ -12,6 +12,7 @@ session = Session()
 
 # Register new medicine
 @medicine.route('/api/v1/Medicine', methods=['POST'])
+@auth.verify_password
 def register():
     # Get data from request body
     data = request.get_json()
@@ -26,6 +27,9 @@ def register():
     exists = session.query(Medicine).filter_by(name=data['name']).first()
     if exists:
         return Response(status=400, response='Medicine with such name already exists.')
+    db_user = session.query(User).filter_by(login=auth.username()).first()
+    if db_user.is_provisor != 1:
+        return Response(status=404, response='Provisor Only')
     # Create new medicine
     new_medicine = Medicine(name=data['name'], manufacturer=data['manufacturer'], price=data['price'], in_stock=True, demand=False, in_stock_number=data['in_stock_number'], demand_number=0)
 
@@ -37,13 +41,17 @@ def register():
 
 
 # Get all medicines
-@medicine.route('/api/v1/Medicine', methods=['GET'])
+@medicine.route('/api/v1/allMedicine', methods=['GET'])
+@auth.verify_password
 def get_reservations():
     # Get all medicines from db
     medicines = session.query(Medicine)
 
     if not medicines:
         return Response(status=404, response='No one medicine')
+    db_user = session.query(User).filter_by(login=auth.username()).first()
+    if db_user.is_provisor != 1:
+        return Response(status=404, response='Provisor Only')
 
     # Return all medicines
     output = []
@@ -59,12 +67,16 @@ def get_reservations():
 
 
 # Get medicine by id
-@medicine.route('/api/v1/medicine/<id_medicine>', methods=['GET'])
+@medicine.route('/api/v1/getmedicine/<id_medicine>', methods=['GET'])
+@auth.verify_password
 def get_user(id_medicine):
     # Check if medicine exists
     db_medicine = session.query(Medicine).filter_by(id_medicine=id_medicine).first()
     if not db_medicine:
         return Response(status=404, response='Medicine with provided ID was not found.')
+    db_user = session.query(User).filter_by(login=auth.username()).first()
+    if db_user.is_provisor != 1:
+        return Response(status=404, response='Provisor Only')
 
     # Return medicine data
     medicine_data = {'id': db_medicine.id_medicine, 'name': db_medicine.name, 'price': db_medicine.price}
@@ -73,6 +85,7 @@ def get_user(id_medicine):
 
 # Update medicine by id
 @medicine.route('/api/v1/medicine/<id_medicine>', methods=['PUT'])
+@auth.verify_password
 def update_user(id_medicine):
     # Get data from request body
     data = request.get_json()
@@ -87,6 +100,9 @@ def update_user(id_medicine):
     db_medicine = session.query(Medicine).filter_by(id_medicine=id_medicine).first()
     if not db_medicine:
         return Response(status=404, response='Medicine with provided ID was not found.')
+    db_user = session.query(User).filter_by(login=auth.username()).first()
+    if db_user.is_provisor != 1:
+        return Response(status=404, response='Provisor Only')
 
     # Check if medicine name is not taken if name already exists
     if 'name' in data.keys():
@@ -120,11 +136,15 @@ def update_user(id_medicine):
 
 # Delete medicine by id
 @medicine.route('/api/v1/medicine/<id_medicine>', methods=['DELETE'])
+@auth.verify_password
 def delete_user(id_medicine):
     # Check if medicine exists
     db_medicine = session.query(Medicine).filter_by(id_medicine=id_medicine).first()
     if not db_medicine:
         return Response(status=404, response='Medicine with provided ID was not found.')
+    db_user = session.query(User).filter_by(login=auth.username()).first()
+    if db_user.is_provisor != 1:
+        return Response(status=404, response='Provisor Only')
 
     # Delete medicine
     session.delete(db_medicine)
